chore: lib/mock-dataを廃止
