+++ packages/app/frontend/lib/today.ts
@@ -1,0 +1,2 @@
+/** モックの「今日」。本番実装では dayjs().format('YYYY-MM-DD') に置き換える */
+export const TODAY = '2026-02-28'
