+++ packages/app/frontend/routes/events/-components/event-card.tsx
@@ -1,6 +1,6 @@
 import { Card, CardContent } from '@frontend/components/ui/card'
 import dayjs from '@frontend/lib/date'
-import { formatCurrency } from '@frontend/lib/mock-data'
+import { formatCurrency } from '@frontend/lib/format'
 import { Link } from '@tanstack/react-router'
 import { ChevronRight } from 'lucide-react'
 import type React from 'react'
